util: add a gsettings key and use that for app_id detection if all else fails
it is very sad that we are resorting to this and i really dislike this piece of code.
lomiri apps (and potentially builds from e.g chum or open repos) are built against a set of toolkits we do not control. for the most part this is fine since most the dependencies are already packaged (or to be packaged) but we don't fork certain things like upstream maintainers do (such as qt)
this leads us down a very dark road. since upstream has essentially custom everything, we have to hack in support for certain things to work [1][2] and this change is yet another hack to make things compatible.
the issue we have here is that appid detection is not exactly standard in any toolkit and system. while gtk and lately qt6 are unifying this, for the most part old qt5 apps either never set the appid or set something that does not represent the app itself.
qmlscene in lomiri is a good example, any app using morph will spawn a process for webapp-container inside qmlscene which then sets its own title and appid, that is besides qmlscene also adding a .qmlscene to the end of appid
then we have cases where the desktop file has the format {developer}.{appname} but the appid says {appname}.{developer} and some even have extra prefixes. in general nothing in consistent
this change will allow the script we generate for each app to set a appid as the fallback for the spawning process using gsettings and phosh will respect that as a last resort if everything else fails (gsettings set io.furios.phosh.shell appid 'org.example.App.desktop').
this key is the reset by phosh itself to not conflict with future apps if they do not set any appid and they also reach a fallback state. this is the cleanest solution i can come up with, with other solutions involving us forking the toolkit and forcing some ugly behaviour that might break apps that are do not come from the open store

[1] https://github.com/FuriLabs/furios-lomiri-app-support
[2] https://github.com/FuriLabs/store-provider/blob/503cfe3850620d9788c833eb07f5525c849b6c85/open_store/click.py#L201

Signed-off-by: Bardia Moshiri <bardia@furilabs.com>